+++ src/librustc_ast/lib.rs
@@ -40,6 +40,7 @@ pub mod util {
 pub mod ast;
 pub mod attr;
 pub use attr::{with_default_globals, with_globals, GLOBALS};
+pub mod crate_disambiguator;
 pub mod entry;
 pub mod expand;
 pub mod mut_visit;
